+++ frontend/models/query/ChatMessageQuery.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace frontend\models;
+namespace frontend\models\query;
 
 /**
  * This is the ActiveQuery class for [[ChatMessage]].
