+++ frontend/models/query/CityQuery.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace frontend\models;
+namespace frontend\models\query;
 
 /**
- * This is the ActiveQuery class for [[Response]].
+ * This is the ActiveQuery class for [[City]].
  *
- * @see Response
+ * @see City
  */
-class ResponseQuery extends \yii\db\ActiveQuery
+class CityQuery extends \yii\db\ActiveQuery
 {
     /*public function active()
     {
@@ -16,7 +16,7 @@ class ResponseQuery extends \yii\db\ActiveQuery
 
     /**
      * {@inheritdoc}
-     * @return Response[]|array
+     * @return City[]|array
      */
     public function all($db = null)
     {
@@ -25,7 +25,7 @@ class ResponseQuery extends \yii\db\ActiveQuery
 
     /**
      * {@inheritdoc}
-     * @return Response|array|null
+     * @return City|array|null
      */
     public function one($db = null)
     {
