+++ frontend/models/query/FavoriteQuery.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace frontend\models;
+namespace frontend\models\query;
 
 /**
- * This is the ActiveQuery class for [[Category]].
+ * This is the ActiveQuery class for [[Favorite]].
  *
- * @see Category
+ * @see Favorite
  */
-class CategoryQuery extends \yii\db\ActiveQuery
+class FavoriteQuery extends \yii\db\ActiveQuery
 {
     /*public function active()
     {
@@ -16,7 +16,7 @@ class CategoryQuery extends \yii\db\ActiveQuery
 
     /**
      * {@inheritdoc}
-     * @return Category[]|array
+     * @return Favorite[]|array
      */
     public function all($db = null)
     {
@@ -25,7 +25,7 @@ class CategoryQuery extends \yii\db\ActiveQuery
 
     /**
      * {@inheritdoc}
-     * @return Category|array|null
+     * @return Favorite|array|null
      */
     public function one($db = null)
     {
